+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.7.2 - February 7, 2023.
 - Added PanoramicToCubemap compute shader, which allows the BlockadeLabsSkyboxGenerator to create cubemaps in a built application.
+- Add XR support with single-pass instanced rendering to the Blockade Depth Skybox shader.
 
 ## 1.7.1 - January 30, 2023.
 - Fix the issue where the `Start Screen` was redisplayed even when disabled in options.